+++ src/renderer.js
@@ -3,47 +3,7 @@
 import ReactFiberReconciler from 'react-reconciler';
 import emptyObject from 'fbjs/lib/emptyObject';
 import { createInstance } from './elements';
-
-const objectsEqual = (a, b) => {
-  const oldPropsKeys = Object.keys(a);
-  const newPropsKeys = Object.keys(b);
-
-  if (oldPropsKeys.length !== newPropsKeys.length) {
-    return false;
-  }
-
-  for (let i = 0; i < oldPropsKeys.length; i++) {
-    const propName = oldPropsKeys[i];
-
-    if (propName === 'render') {
-      if (!a[propName] !== !b[propName]) {
-        return false;
-      }
-      continue;
-    }
-
-    if (propName !== 'children' && a[propName] !== b[propName]) {
-      if (
-        typeof a[propName] === 'object' &&
-        typeof b[propName] === 'object' &&
-        objectsEqual(a[propName], b[propName])
-      ) {
-        continue;
-      }
-
-      return false;
-    }
-
-    if (
-      propName === 'children' &&
-      (typeof a[propName] === 'string' || typeof b[propName] === 'string')
-    ) {
-      return a[propName] === b[propName];
-    }
-  }
-
-  return true;
-};
+import propsEqual from './utils/propsEqual';
 
 // If the Link has a strign child or render prop, substitute the instance by a Text,
 // that will ultimately render the inline Link via the textkit PDF renderer.
@@ -87,7 +47,7 @@ const PDFRenderer = ReactFiberReconciler({
   },
 
   prepareUpdate(element, type, oldProps, newProps) {
-    return !objectsEqual(oldProps, newProps);
+    return !propsEqual(oldProps, newProps);
   },
 
   resetAfterCommit() {
